* Initial Commit

* Initial Commit

+++ lib/configChecker.ts
@@ -18,6 +18,7 @@ export const checkConfigFile = (): Config | null => {
         const filePath = path.resolve(process.cwd(), configFile);
 
         if (fs.existsSync(filePath)) {
+            console.log(`Configuration file (${configFile}) already exists in the root.`);
             return require(filePath) as Config;
         }
     }
@@ -26,15 +27,19 @@ export const checkConfigFile = (): Config | null => {
 };
 
 export const initializeConfigFile = (): void => {
-    const defaultConfig: Config = {
-        router: {
-            includes: ["./**/*.{ts,js,tsx,jsx}"],
-            excludes: [],
-        },
-    };
-
     const configFilePath = path.resolve(process.cwd(), 'nextrouterkit.config.json');
-    fs.writeFileSync(configFilePath, JSON.stringify(defaultConfig, null, 2));
 
-    console.log('nextrouterkit.config.json created successfully with default configuration.');
+    if (fs.existsSync(configFilePath)) {
+        console.log('Configuration file already exists. No action taken.');
+    } else {
+        const defaultConfig: Config = {
+            router: {
+                includes: ["./**/*.{ts,js,tsx,jsx}"],
+                excludes: [],
+            },
+        };
+
+        fs.writeFileSync(configFilePath, JSON.stringify(defaultConfig, null, 2));
+        console.log('nextrouterkit.config.json created successfully with default configuration.');
+    }
 };
